+++ ebs-20210730/ChangeLog.md
@@ -1,3 +1,10 @@
+2022-05-27 Version: 1.2.1
+- Add describe dbsc for ecs api.
+- Add disk replica pair api.
+- Add disk replica group api.
+- Bug fix.
+- SourceRegionId change to RegionId.
+
 2022-04-12 Version: 1.1.0
 - Add disk replica pair api.
 - Add disk replica group api.